Remove DirectoryWrapper
