+++ lib/builder.js
@@ -143,13 +143,8 @@ module.exports = class Builder extends EventEmitter {
         }),
       () => {
         this._cancelationRequest = null;
-        this._settleNodesState();
       }
     );
-  }
-
-  _settleNodesState() {
-    this.nodeWrappers.forEach(nw => nw.settle());
   }
 
   cancel() {
