+++ lib/wrappers/node.js
@@ -5,6 +5,27 @@ const undefinedToNull = require('../utils/undefined-to-null');
 module.exports = class NodeWrapper {
   constructor() {
     this.buildState = {};
+    this._revision = 0;
+    this._revised = false;
+  }
+
+  revise() {
+    this._revised = true;
+  }
+
+  get revision() {
+    return this._revision;
+  }
+
+  set revision(value) {
+    this._revision = value;
+  }
+
+  settle() {
+    if (this._revised) {
+      this._revised = false;
+      this._revision++;
+    }
   }
 
   toJSON() {
