+++ lib/wrappers/source-node.js
@@ -1,14 +1,12 @@
 'use strict';
 
 const NodeWrapper = require('./node');
-const DirectoryWrapper = require('./directory');
 const fs = require('fs');
 const undefinedToNull = require('../utils/undefined-to-null');
 
 module.exports = class SourceNodeWrapper extends NodeWrapper {
   setup(/* features */) {
-    this.outputWrapper = new DirectoryWrapper(this.nodeInfo.sourceDirectory);
-    this.outputWrapper.revise(); // set initial revised state to trigger first build
+    this.revise(); // initialize with a "change" to kick off the build
   }
   build() {
     // We only check here that the sourceDirectory exists and is a directory
@@ -27,14 +25,7 @@ module.exports = class SourceNodeWrapper extends NodeWrapper {
 
     this.buildState.selfTime = 0;
     this.buildState.totalTime = 0;
-  }
-
-  revise() {
-    this.outputWrapper.revise();
-  }
-
-  settle() {
-    this.outputWrapper.settle();
+    this.settle();
   }
 
   toString() {
