+++ lib/wrappers/transform-node.js
@@ -1,7 +1,6 @@
 'use strict';
 
 const NodeWrapper = require('./node');
-const DirectoryWrapper = require('./directory');
 const fs = require('fs');
 const undefinedToNull = require('../utils/undefined-to-null');
 const rimraf = require('rimraf');
@@ -14,20 +13,24 @@ module.exports = class TransformNodeWrapper extends NodeWrapper {
       cachePath: this.cachePath,
     });
     this.callbackObject = this.nodeInfo.getCallbackObject();
-    this.outputWrapper = new DirectoryWrapper(this.outputPath);
     this.callbackObject.revise = () => {
-      this.outputWrapper.revise();
+      this.revise();
     };
   }
 
-  shouldBuild() {
+  getInputsMaxRevision() {
+    let max = this.revision;
     for (let inputNodeWrapper of this.inputNodeWrappers) {
-      if (inputNodeWrapper.outputWrapper.hasChanges()) {
-        return true;
+      if (inputNodeWrapper.revision > max) {
+        max = inputNodeWrapper.revision;
       }
     }
 
-    return false;
+    return max;
+  }
+
+  shouldBuild() {
+    return this.getInputsMaxRevision() > this.revision;
   }
 
   build() {
@@ -55,11 +58,8 @@ module.exports = class TransformNodeWrapper extends NodeWrapper {
       for (let i = 0; i < this.inputNodeWrappers.length; i++) {
         this.buildState.totalTime += this.inputNodeWrappers[i].buildState.totalTime;
       }
+      this.settle();
     });
-  }
-
-  settle() {
-    this.outputWrapper.settle();
   }
 
   toString() {
